+++ src/Hyperledger.Aries.AspNetCore.Contracts/Features/Connection/CreateInvitation/CreateInvitationRequest.cs
@@ -1,7 +1,7 @@
 namespace Hyperledger.Aries.AspNetCore.Features.Connections
 {
-  using Hyperledger.Aries.AspNetCore.Features.Bases;
-  using Hyperledger.Aries.Features.DidExchange;
+  using Aries.Features.Handshakes.Connection.Models;
+  using Bases;
   using MediatR;
 
   /// <summary>
