+++ src/Hyperledger.Aries.Routing.Mediator/MediatorProvisioningService.cs
@@ -3,8 +3,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Configuration;
 using Hyperledger.Aries.Extensions;
-using Hyperledger.Aries.Features.DidExchange;
 using Hyperledger.Aries.Features.Handshakes.Connection;
+using Hyperledger.Aries.Features.Handshakes.Connection.Models;
 using Hyperledger.Aries.Storage;
 using Hyperledger.Indy.WalletApi;
 using Microsoft.Extensions.Hosting;
