+++ src/Hyperledger.Aries.TestHarness/Mock/InProcAgent.cs
@@ -5,9 +5,9 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Agents;
-using Hyperledger.Aries.Features.DidExchange;
 using Hyperledger.Aries.Features.Handshakes.Common;
 using Hyperledger.Aries.Features.Handshakes.Connection;
+using Hyperledger.Aries.Features.Handshakes.Connection.Models;
 using Hyperledger.Aries.Payments;
 using Hyperledger.Aries.Routing;
 using Hyperledger.Aries.Routing.Mediator.Storage;
