+++ src/Hyperledger.Aries/Features/Handshakes/Connection/DefaultConnectionService.cs
@@ -12,9 +12,9 @@ using Hyperledger.Aries.Decorators.Threading;
 using Hyperledger.Aries.Extensions;
 using Hyperledger.Aries.Features.Handshakes.Common;
 using Hyperledger.Aries.Features.Handshakes.Common.Dids;
+using Hyperledger.Aries.Features.Handshakes.Connection.Extensions;
 using Hyperledger.Aries.Features.Handshakes.Connection.Models;
 using Hyperledger.Aries.Models.Events;
-using Hyperledger.Aries.Features.DidExchange.Models;
 using Hyperledger.Aries.Storage;
 using Hyperledger.Aries.Utils;
 using Hyperledger.Indy.CryptoApi;
