+++ src/Hyperledger.Aries/Features/IssueCredential/Abstractions/ICredentialService.cs
@@ -1,7 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Agents;
+using Hyperledger.Aries.Common;
 using Hyperledger.Aries.Features.Handshakes.Common;
+using Hyperledger.Aries.Features.IssueCredential.Models;
 using Hyperledger.Aries.Storage;
 
 namespace Hyperledger.Aries.Features.IssueCredential
