+++ test/Hyperledger.Aries.Tests/Protocols/ConnectionTests.cs
@@ -4,10 +4,8 @@ using System.Linq;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Agents;
-using Hyperledger.Aries.Models.Events;
 using Hyperledger.Aries.Configuration;
 using Hyperledger.Aries.Contracts;
-using Hyperledger.Aries.Features.DidExchange;
 using Hyperledger.Aries.Features.Handshakes.Common;
 using Hyperledger.Aries.Features.Handshakes.Common.Dids;
 using Hyperledger.Aries.Features.Handshakes.Connection;
