+++ test/Hyperledger.Aries.Tests/Protocols/CredentialTests.cs
@@ -9,8 +9,8 @@ using Hyperledger.Aries.Agents;
 using Hyperledger.Aries.Configuration;
 using Hyperledger.Aries.Contracts;
 using Hyperledger.Aries.Extensions;
-using Hyperledger.Aries.Features.DidExchange;
 using Hyperledger.Aries.Features.Handshakes.Connection;
+using Hyperledger.Aries.Features.Handshakes.Connection.Models;
 using Hyperledger.Aries.Features.IssueCredential;
 using Hyperledger.Aries.Ledger;
 using Hyperledger.Aries.Models.Events;
